refactor(auth): fold the floxhub-auth crate into floxhub-client as the auth module
Review feedback: auth belongs with the client it authenticates. The
module keeps the one-file-per-type layout and stays transport-free;
lib.rs re-exports auth::* so downstream imports are unchanged.
