--- cli/floxhub-auth/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "floxhub-auth"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-chrono.workspace = true
-jsonwebtoken.workspace = true
-serde.workspace = true
-serde_with.workspace = true
-thiserror.workspace = true
-tracing.workspace = true
-url.workspace = true
-
-# GSSAPI/Kerberos authentication
-libgssapi.workspace = true
-base64.workspace = true
