+++ cli/floxhub-client/Cargo.toml
@@ -6,15 +6,17 @@ edition.workspace = true
 [dependencies]
 async-stream.workspace = true
 catalog-api-v1.workspace = true
+chrono.workspace = true
 factory-api-v1.workspace = true
 derive_more.workspace = true
-floxhub-auth.workspace = true
 futures.workspace = true
 http.workspace = true
+jsonwebtoken.workspace = true
 reqwest.workspace = true
 sentry.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+serde_with.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
@@ -23,9 +25,12 @@ url.workspace = true
 httpmock.workspace = true
 serde_yaml.workspace = true
 
+# GSSAPI/Kerberos authentication
+libgssapi.workspace = true
+base64.workspace = true
+
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
-chrono.workspace = true
 serde_json.workspace = true
 itertools.workspace = true
 proptest.workspace = true
